+++ heroku/database.py
@@ -116,7 +116,7 @@ class Database(dict):
                 "heroku-assets",
                 "🌆 Your Heroku assets will be stored here",
                 archive=True,
-                avatar="https://raw.githubusercontent.com/coddrago/Heroku/dev-test/assets/heroku-assets.png"
+                avatar="https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/heroku_assets.png"
             )
         except ChannelsTooMuchError:
             self._assets = None