+++ heroku/modules/help.py
@@ -333,7 +333,7 @@ class Help(loader.Module):
 
         await utils.answer(
             message,
-            (self.config["desc_icon"] + " {}\n <blockquote exp>{}</blockquote><blockquote exp>{}</blockquote><blockquote exp>{}</blockquote>").format(
+            (self.config["desc_icon"] + " {}\n <blockquote expandable>{}</blockquote><blockquote expandable>{}</blockquote><blockquote expandable>{}</blockquote>").format(
                 reply,
                 "".join(core_),
                 "".join(plain_ + (no_commands_ if force else [])),