+++ heroku/modules/inline_stuff.py
@@ -18,8 +18,7 @@ from herokutl.tl.functions.contacts import UnblockRequest
 from herokutl.tl.types import Message
 
 from .. import loader, utils
-from ..inline.types import BotInlineMessage
-
+from ..inline.types import BotInlineMessage, InlineCall
 
 @loader.tds
 class InlineStuff(loader.Module):
@@ -125,10 +124,47 @@ class InlineStuff(loader.Module):
         await utils.answer(message, self.strings("bot_updated"))
 
     async def aiogram_watcher(self, message: BotInlineMessage):
-        if message.text != "/start":
+        if message.text != "/start" and message.text != "/profile":
             return
 
-        await message.answer_photo(
-            "https://imgur.com/a/0gmlFYI.png",
-            caption=self.strings("this_is_heroku"),
-        )
+        if message.text == "/start":
+            await message.answer_photo(
+                "https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/start_cmd.png",
+                caption=self.strings("this_is_heroku"),
+            )
+
+        if message.text == "/profile":
+            
+            if message.from_user.id != self.client.tg_id:
+                await message.answer("❌ You are not allowed to use this")
+            else:
+                await message.answer_photo(
+                    "https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/start_cmd.png",
+                    caption = self.strings["profile_cmd"].format(prefix=self.get_prefix(),ram_usage=utils.get_ram_usage(),cpu_usage=utils.get_cpu_usage(),host=utils.get_named_platform()), 
+                    reply_markup = self.inline.generate_markup(
+                        markup_obj=[
+                            [
+                                {
+                                    "text": "🚀 Restart", 
+                                    "callback": self.restart, 
+                                    "args": (message,)
+                                }
+                            ],
+                            [
+                                {
+                                    "text": "⚠️ Reset prefix", 
+                                    "callback": self.reset_prefix,
+                                    "args": (message,)
+                                }
+                            ]
+                        ]
+                    )
+                )
+
+    async def restart(self, call: InlineCall, message):
+        await call.edit(self.strings["restart"])
+        await self.invoke("restart", "-f", message=message, peer=self.inline.bot.id)
+
+    async def reset_prefix(self, call: InlineCall, message):
+        await message.answer(self.strings["prefix_reset"])
+        self.db.set("heroku.main", "command_prefix", ".")