+++ heroku/modules/presets.py
@@ -123,7 +123,7 @@ class Presets(loader.Module):
     async def _menu(self):
         await self.inline.bot.send_photo(
             self._client.tg_id,
-            'https://imgur.com/a/7xwuOEt.png',
+            'https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/presets_cmd.png',
             caption=self.strings('welcome'),
             reply_markup=self.inline.generate_markup(self._markup),
         )