+++ heroku/modules/settings.py
@@ -108,9 +108,8 @@ class CoreMod(loader.Module):
 
     @loader.command(ru_doc="Информация о Хероку", en_doc="Information of Heroku", ua_doc="Інформація про Хероку", de_doc="Informationen über Heroku")
     async def herokucmd(self, message: Message):
-        await utils.answer_file(
+        await utils.answer(
             message,
-            "https://imgur.com/a/i0Mq22X.png",
             self.strings("heroku").format(
                 (
                     utils.get_platform_emoji()
@@ -126,6 +125,8 @@ class CoreMod(loader.Module):
                 if version.branch == "master"
                 else self.strings("unstable").format(version.branch)
             ),
+            file= "https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/heroku_cmd.png",
+            reply_to=getattr(message, "reply_to_msg_id", None),
         )
 
     @loader.command()
@@ -319,14 +320,14 @@ class CoreMod(loader.Module):
                 self.strings("choose_installation"),
                 message,
                 reply_markup=self._markup,
-                photo="https://imgur.com/a/HrrFair.png",
+                photo="https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/heroku_installation.png",
                 disable_security=True
         )
             ):
 
             await self.client.send_file(
                 message.peer_id,
-                "https://imgur.com/a/HrrFair.png",
+                "https://raw.githubusercontent.com/coddrago/assets/refs/heads/main/heroku/heroku_installation.png",
                 caption=self.strings["installation"], reply_to=getattr(message, "reply_to_msg_id", None),)
         elif "-v" in args:
             await utils.answer(message, self.strings["vds_install"])